requirements.txt
django
aiogram
pillow
numpy 
noise